Make the baseline attack a separate file + inherits from AttributeInferenceAttack
